Continue refactoring to use new path/name
