+++ browser/src/Plugins/Api/Process.ts
@@ -1,7 +1,7 @@
 import * as ChildProcess from "child_process"
 
-import * as Config from "./../../Config"
 import * as Platform from "./../../Platform"
+import { configuration } from "./../../Services/Configuration"
 
 const getPathSeparator = () => {
     return Platform.isWindows() ? ";" : ":"
@@ -29,7 +29,7 @@ const mergeSpawnOptions = (originalSpawnOptions: ChildProcess.ExecOptions | Chil
 
     const existingPath = process.env.Path || process.env.PATH
 
-    requiredOptions.env.PATH = mergePathEnvironmentVariable(existingPath, Config.instance().getValue("environment.additionalPaths"))
+    requiredOptions.env.PATH = mergePathEnvironmentVariable(existingPath, configuration.getValue("environment.additionalPaths"))
 
     return {
         ...originalSpawnOptions,
